New WaitForSynchronizationAsync method (and hints for Long/Short running commands).
This ensures that "short running commands" are done.
Long running commands are the ones that are not completed once
handled (they are completed in the background). For them, one idea
could be to add a Task<bool> IsLongRunningCommand on the command.
Completion will TrySet to true and, at the end of the HandleCommand, we
TrySet to false. Thanks to this, one can await the handling of the command,
and/or its completion.

This enables Long vs. Short running commands choice to be "dynamic".

For Cris, this de facto implements the 'A' vs. 'S' VISA status for
background/remote command handler...